Major code decoupling
- Decoupled previous report controller and multiple service calls
- Added controller and service for generating PDF reports
- Renamed entity fields for report/medical parameters (adjusted dummy script for DB)
- Added reportType for reports and removed hardcoded present logic of generating data
- Renamed packages with capital first letters
