--- src/main/java/org/med/Entity/PromptEntity.java
@@ -1,21 +1,0 @@
-package org.med.Entity;
-
-import io.smallrye.common.constraint.NotNull;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "prompt")
-public class PromptEntity {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @NotNull
-    private String description;
-    @NotNull
-    private String promptValue;
-}
