--- src/main/java/org/med/Validator/ValidBloodType.java
@@ -1,20 +1,0 @@
-package org.med.Validator;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.FIELD})
-@Constraint(validatedBy = BloodTypeValidator.class)
-public @interface ValidBloodType {
-    String message() default "Invalid value";
-
-    Class<?>[] groups() default {};
-
-    Class<? extends Payload>[] payload() default {};
-}
