+++ src/main/java/org/med/entity/PatientEntity.java
@@ -1,0 +1,33 @@
+package org.med.entity;
+
+import io.smallrye.common.constraint.NotNull;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.med.validator.ValidBloodType;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Builder
+@Table(name = "patient")
+public class PatientEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @NotNull
+    private String firstName;
+    @NotNull
+    private String lastName;
+    @ValidBloodType
+    private String bloodType;
+    @NotNull
+    private Character sex;
+    @NotNull
+    private Integer age;
+
+    private Float weight;
+}
