+++ src/main/java/org/med/repository/MedicalParameterRepository.java
@@ -1,0 +1,11 @@
+package org.med.repository;
+
+import io.quarkus.hibernate.orm.panache.PanacheRepository;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.transaction.Transactional;
+import org.med.entity.MedicalParameterEntity;
+
+@ApplicationScoped
+@Transactional
+public class MedicalParameterRepository implements PanacheRepository<MedicalParameterEntity> {
+}
