+++ src/main/java/org/med/repository/PatientRepository.java
@@ -1,0 +1,9 @@
+package org.med.repository;
+
+import io.quarkus.hibernate.orm.panache.PanacheRepository;
+import jakarta.enterprise.context.ApplicationScoped;
+import org.med.entity.PatientEntity;
+
+@ApplicationScoped
+public class PatientRepository implements PanacheRepository<PatientEntity> {
+}
