+++ src/main/java/org/med/validator/ValidBloodType.java
@@ -1,0 +1,20 @@
+package org.med.validator;
+
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.FIELD})
+@Constraint(validatedBy = BloodTypeValidator.class)
+public @interface ValidBloodType {
+    String message() default "Invalid value";
+
+    Class<?>[] groups() default {};
+
+    Class<? extends Payload>[] payload() default {};
+}
