+++ src/main/resources/scripts/loadDummyData.sql
@@ -10,10 +10,10 @@ ALTER TABLE patient ALTER COLUMN weight DROP NOT NULL;
 INSERT INTO patient(firstName, lastName, bloodType, sex, age) VALUES ('John', 'Doe', 'A+', 'M', 45);
 
 -- Report data dummy
-INSERT INTO report (patient_id) values (1);
+INSERT INTO report (patient, reportType) values (1, 'Blood');
 
 -- Medical Parameter dummy data
-INSERT INTO medical_parameter(parameterName, report_id, parameterValue, referenceValueRange) VALUES ('Hemoglobin', 1, '132 g/L', '135 – 175 g/L (men), 120 – 160 g/L (women)');
-INSERT INTO medical_parameter(parameterName, report_id, parameterValue, referenceValueRange) VALUES ('White Blood Cells', 1, '7.2 x10⁹/L', '4.0 – 10.0 x10⁹/L');
-INSERT INTO medical_parameter(parameterName, report_id, parameterValue, referenceValueRange) VALUES ('Hematocrit', 1, '0.41 L/L', ' 0.40 – 0.50 L/L (men), 0.35 – 0.47 L/L (women)');
-INSERT INTO medical_parameter(parameterName, report_id, parameterValue) VALUES ('Red Cell Distribution Width (RDW)', 1, '15.5%');
+INSERT INTO medical_parameter(parameterName, report, parameterValue, referenceValueRange) VALUES ('Hemoglobin', 1, '132 g/L', '135 – 175 g/L (men), 120 – 160 g/L (women)');
+INSERT INTO medical_parameter(parameterName, report, parameterValue, referenceValueRange) VALUES ('White Blood Cells', 1, '7.2 x10⁹/L', '4.0 – 10.0 x10⁹/L');
+INSERT INTO medical_parameter(parameterName, report, parameterValue, referenceValueRange) VALUES ('Hematocrit', 1, '0.41 L/L', ' 0.40 – 0.50 L/L (men), 0.35 – 0.47 L/L (women)');
+INSERT INTO medical_parameter(parameterName, report, parameterValue) VALUES ('Red Cell Distribution Width (RDW)', 1, '15.5%');
