+++ src/test/java/org/med/utils/DBCleanUtil.java
@@ -1,0 +1,26 @@
+package org.med.utils;
+
+import jakarta.inject.Inject;
+import jakarta.inject.Singleton;
+import org.med.service.MedicalParameterService;
+import org.med.service.PatientService;
+import org.med.service.PromptService;
+import org.med.service.ReportService;
+@Singleton
+public class DBCleanUtil {
+    @Inject
+    private PatientService patientService;
+    @Inject
+    private ReportService reportService;
+    @Inject
+    private MedicalParameterService medicalParameterService;
+    @Inject
+    private PromptService promptService;
+
+    public void cleanDB() {
+        reportService.deleteAll();
+        medicalParameterService.deleteAll();
+        patientService.deleteAll();
+        promptService.deleteAll();
+    }
+}
